setValue added to OutputPort. Created unique folder and files for each Node
