+++ app/src/main/java/Model/Nodes/XORNode.java
@@ -1,0 +1,21 @@
+package Model.Nodes;
+
+/**
+ * Implements the logical XOR gate. The gate is true if exactly 1 input is true.
+ * 
+ * @author victh
+ *
+ */
+public class XORNode extends Model.Node {
+	public XORNode(String tag) {
+		super(tag);
+	}
+	public void evaluate() {
+		if(inputs.stream().filter(x -> x.getValue() == true).count() == 1) {
+			setOutputs(true);
+		}
+		else {
+			setOutputs(false);
+		}
+	}
+}
